disabled.feature rebase_onto_remove_deleted.go

+++ internal/vm/opcodes/rebase_onto_remove_deleted.go
@@ -39,12 +39,14 @@ func (self *RebaseOntoRemoveDeleted) Run(args shared.RunArgs) error {
 		if err != nil {
 			return fmt.Errorf("cannot determine conflicting files after rebase: %w", err)
 		}
-		for _, conflictingFile := range conflictingFiles {
-			if conflictingChange, has := conflictingFile.CurrentBranchChange.Get(); has {
-				_ = args.Git.ResolveConflict(args.Frontend, conflictingChange.FilePath, gitdomain.ConflictResolutionTheirs)
-				_ = args.Git.StageFiles(args.Frontend, conflictingChange.FilePath)
-			} else if baseChange, has := conflictingFile.BaseChange.Get(); has {
-				_ = args.Git.RemoveFile(args.Frontend, baseChange.FilePath)
+		if args.Config.Value.NormalConfig.AutoResolve.ShouldAutoResolve() {
+			for _, conflictingFile := range conflictingFiles {
+				if conflictingChange, has := conflictingFile.CurrentBranchChange.Get(); has {
+					_ = args.Git.ResolveConflict(args.Frontend, conflictingChange.FilePath, gitdomain.ConflictResolutionTheirs)
+					_ = args.Git.StageFiles(args.Frontend, conflictingChange.FilePath)
+				} else if baseChange, has := conflictingFile.BaseChange.Get(); has {
+					_ = args.Git.RemoveFile(args.Frontend, baseChange.FilePath)
+				}
 			}
 		}
 		_ = args.Git.ContinueRebase(args.Frontend)
